feat: Create Order Controller and Service, Dtos for OrderItem, add phone field for user, fix roles and protected routes,

+++ src/main/java/com/saihoz/e_commerce_api/order/OrderItem.java
@@ -7,7 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -31,5 +30,5 @@ public class OrderItem {
 
     private Integer quantity;
 
-    private BigDecimal unitPrice;
+    private double unitPrice;
 }
